split the agnostic retirement tests among 4/5 files
